Synchronize events from google API Calendar
