+++ app/src/main/java/com/framgia/youralarm1/contstant/Const.java
@@ -47,5 +47,5 @@ public class Const {
     public static final String SATUDAY = "SA";
     public static final String DELETE_FAIL = "fail";
     public static final String DELETE_SUCCESS = "success";
-
+    public static final String DEFAULT_EVENT_ID = "deefault_event_id";
 }
