+++ app/src/main/java/com/framgia/youralarm1/util/DeleteEvent.java
@@ -1,0 +1,85 @@
+package com.framgia.youralarm1.util;
+
+import android.app.Activity;
+import android.app.ProgressDialog;
+import android.content.Context;
+import android.os.AsyncTask;
+import android.widget.Toast;
+
+import com.framgia.youralarm1.R;
+import com.framgia.youralarm1.contstant.Const;
+import com.framgia.youralarm1.models.ItemAlarm;
+import com.google.api.client.extensions.android.http.AndroidHttp;
+import com.google.api.client.googleapis.extensions.android.gms.auth.GoogleAccountCredential;
+import com.google.api.client.http.HttpTransport;
+import com.google.api.client.json.JsonFactory;
+import com.google.api.client.json.jackson2.JacksonFactory;
+
+import java.io.IOException;
+
+/**
+ * Created by phongtran on 16/06/2016.
+ */
+public class DeleteEvent extends AsyncTask<Void, Void, String> {
+    private Activity mContext;
+    private com.google.api.services.calendar.Calendar mService = null;
+    private ItemAlarm mItemAlarm;
+    private ProgressDialog mProgress;
+    private GoogleAccountCredential mCredential;
+
+    public DeleteEvent(Activity context, GoogleAccountCredential credential, ItemAlarm itemAlarm) {
+        mContext = context;
+        mItemAlarm = itemAlarm;
+        mCredential = credential;
+        HttpTransport transport = AndroidHttp.newCompatibleTransport();
+        JsonFactory jsonFactory = JacksonFactory.getDefaultInstance();
+        mService = new com.google.api.services.calendar.Calendar.Builder(
+                transport, jsonFactory, credential)
+                .setApplicationName(mContext.getResources().getString(R.string.application_name))
+                .build();
+    }
+
+    @Override
+    protected void onPreExecute() {
+        super.onPreExecute();
+        if (mItemAlarm.getIdEvent().equals("")) {
+            cancel(true);
+        }
+        mProgress = new ProgressDialog(mContext);
+        mProgress.setMessage(mContext.getString(R.string.delete_api));
+        mProgress.show();
+    }
+
+    @Override
+    protected String doInBackground(Void... params) {
+        try {
+            mService.events().delete(mContext.getResources()
+                    .getString(R.string.primary), mItemAlarm.getIdEvent()).execute();
+            return Const.DELETE_SUCCESS;
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return Const.DELETE_FAIL;
+    }
+
+    @Override
+    protected void onPostExecute(String aVoid) {
+        if (mProgress != null && mProgress.isShowing()) {
+            mProgress.dismiss();
+        }
+        if (aVoid.equals(Const.DELETE_SUCCESS)) {
+            new MakeRequestTask(mContext, mCredential).execute();
+        } else {
+            Toast.makeText(mContext, R.string.delete_fail, Toast.LENGTH_SHORT).show();
+        }
+    }
+
+    @Override
+    protected void onCancelled() {
+        super.onCancelled();
+        if (mProgress != null && mProgress.isShowing()) {
+            mProgress.dismiss();
+        }
+        Toast.makeText(mContext, R.string.delete_fail, Toast.LENGTH_SHORT).show();
+    }
+}
